requirements.txt
ansicon==1.89.0
anyio==4.5.2
astunparse==1.6.3
blessed==1.20.0
botocore==1.35.44
certifi==2024.8.30
chalice==1.29.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
graphql-core==3.2.5
greenlet==3.1.1
h11==0.14.0
idna==3.10
inquirer==2.10.1
jinxed==1.3.0
jmespath==1.0.1
libcst==1.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
mypy-extensions==1.0.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-editor==1.0.4
python-multipart==0.0.12
PyYAML==6.0.2
readchar==4.2.0
requests==2.32.3
rich==13.9.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.0
strawberry-graphql==0.246.2
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.32.0
wcwidth==0.2.13